cover'n'peaking
When considering cover the AI now also takes into account what direction it wants the cover from. Not much value in standing next to a wall, when all enemies are on the same side of it.

Cover has more important in the decision-making.

Various tweaks and fixes about peaking-behaviour.

+++ src/Battlescape/AIModule.h
@@ -219,7 +219,7 @@ public:
 	/// Non-cheating-AI needs to be able to determine whether the enemy is doing Triton-shenanigans, where we should prevent exposing ourselves or is exposed enough themselves for us to strike
 	bool IsEnemyExposedEnough();
 	/// Get the cover-value of a tile
-	int getCoverValue(Tile *tile, BattleUnit *bu);
+	float getCoverValue(Tile *tile, BattleUnit *bu);
 	/// checks whethere there's any cover in range
 	bool anyCoverInRange(const std::vector<PathfindingNode *> nodeVector);
 };
